+++ src/push_swap/move_to_stack_b.c
@@ -6,7 +6,7 @@
 /*   By: yquaro <yquaro@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/25 16:26:12 by yquaro            #+#    #+#             */
-/*   Updated: 2020/01/20 14:12:53 by yquaro           ###   ########.fr       */
+/*   Updated: 2020/02/07 22:48:02 by yquaro           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ static size_t			chunk_element_from_top(t_stack *stack, size_t chunk_num)
 	i = 0;
 	while (i < stack->a->used_size)
 	{
-		if (!is_fixed(stack, i) && stack->a->num[i] <= BORDER)
+		if (!is_fixed(stack, i) && stack->a->num[i] <= (int)BORDER)
 			return (i);
 		i++;
 	}
@@ -51,7 +51,7 @@ static size_t			chunk_element_from_down(t_stack *stack, \
 	i = stack->a->used_size - 1;
 	while (i)
 	{
-		if (!is_fixed(stack, i) && stack->a->num[i] <= BORDER)
+		if (!is_fixed(stack, i) && stack->a->num[i] <= (int)BORDER)
 			return (i);
 		i--;
 	}
